Fixed bad SHA1 bug, and work with good SHA1 hashes. Merged star office into libre #3089.
